Add view and procedure to handle missing dimension slices
Due to a bug in versions prior to 1.7.5, dimension slices could be
removed when running insert and `drop_chunks` concurrently. This commit
contain a script to re-construct such missing dimension slices as well
as a view to show any missing slices and what hypertables and chunks
they affect.

See timescale/timescaledb#1986