Deal with non-existent chunks
Before this, rows in the source table could have been missed if they
fall outside chunk boundaries (ie if their insertion would cause a new
chunk to be created in the hypertable). This solves this by also
inserting into any gaps in the chunks as well as into the beginning or
end where the chunks may not exist yet.