Fix backfill.sql compression job rescheduling
In `move_compression_job()`, the existing implementation was
`INNER JOIN`ing  on `timescaledb_information.jobs` and
`timescaledb_information.job_stats`.

This was both redundant, (since we were not using any information from
the `job_stats` view, and also prone to failure, since a job that had
never run would have no entry in the `job_stats` view.

This would cause the job not to be rescheduled, causing it to be run at
the same time as the backfill proceedure, leading to data corruption.

---

Secondly, if the Compression Policy job was already running when the
jobs table was examined, the `next_start` would be Timescale's
`DT_NOBEGIN` (sometimes displayed as `-infinity`), which would cause the
reschedule call at the end of the backfill to fail.

By sleeping in a loop until the job is no longer runnning, we avoid this
edge case. The majority of the time, no sleep is required however.

---

Some `RAISE NOTICE` statements have been included to make future
debugging of this kind easier, and providing more visibility into the
actions the backfill proceedure is taking.

Co-authored-by: Simon Schmidt <simon.schmidt@infogrid.io>
Co-authored-by: Nick Pope <nick.pope@infogrid.io>