Script to fix invalidation threshold and materialization log (#80)
For some reasons, there are services with some hypertable having
invalidation near infinity, same for the lowest_modified_value in
materialization_invalidation_log. We have not managed to trace down
the root cause, so we plan to run the script on each of the affected
service to fix the data, then will monitor if this happens again.

This is to be run manually on the affected services